pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auth-service"
version = "0.1.0"
description = "DDD/CQRS Authentication Service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "strawberry-graphql[fastapi]>=0.243.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "bcrypt>=4.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.9.0",
    "structlog>=24.0.0",
    "redis[asyncio]>=5.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "anyio>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["auth_service*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "module"
asyncio_default_test_loop_scope = "module"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require running services)",
    "smoke: marks tests as smoke tests (require running infrastructure)",
]

[tool.ruff]
line-length = 100
target-version = "py312"
